Move app_template_demo.py -> demos/
