+++ hyperdiv_docs/demos/app_template_demo.py
